edit device by Brandi

+++ src/sections/device/ListItem.js
@@ -1,33 +1,79 @@
-import React from 'react';
-import {View, Text, StyleSheet, TouchableOpacity} from 'react-native';
-import {connect} from 'react-redux';
-import {deleteDevice} from '../../../Actions';
+import React, { useState } from 'react';
+import { View, Text, StyleSheet, TouchableOpacity, Alert } from 'react-native';
+import { Button, Overlay, Input } from 'react-native-elements';
+
+import { connect } from 'react-redux';
+import { deleteDevice, editDevice } from '../../../Actions';
 import Icon from '../../utils/Icon';
-const ListItem = ({item, theme, navigation, deleteDevice}) => {
+
+const ListItem = ({ item, theme, navigation, deleteDevice, editDevice, screen }) => {
   const width = 20;
   const height = 24;
-  const handleDeleteDevice = () => {
-    deleteDevice({
-      phoneNumber: item.phoneNumber,
+
+  const [state, set] = useState({
+      visible: false,
+      phoneEdit: '',
+      nameEdit: ''
     });
+
+    
+    
+  const handleDeleteDevice = () => {
+    
+    
+    Alert.alert(
+      screen.alerts.delete,
+      screen.alerts.deleteAsk,
+      [
+        {
+          text: screen.alerts.add_cancel_label,
+          onPress: () => {
+            console.log('Cancel')
+          }
+        },
+        {
+          text: screen.alerts.add_confirm_label,
+          onPress: () => {
+            deleteDevice({
+              phoneNumber: item.phoneNumber,
+            });
+          }
+        }
+      ]
+
+    )
+  };
+  const handleEditDevice = () => {
+    editDevice({
+      phoneNumber: item.phoneNumber,
+      name:item.name,
+      phoneEdit: state.phoneEdit,
+      nameEdit: state.nameEdit
+    });
+    toggleOverlay();
+  }
+
+ 
+  const toggleOverlay = () => {
+    set({ ...state, visible: !state.visible });
   };
   return (
-    <View style={[style.container, {borderColor: theme.device_list_border}]}>
+    <View style={[style.container, { borderColor: theme.device_list_border }]}>
       <TouchableOpacity
         onPress={() =>
-          navigation.navigate('DeviceControl', {phone: item.phoneNumber})
+          navigation.navigate('DeviceControl', { phone: item.phoneNumber })
         }
         style={[
           style.picture_name_container,
-          {backgroundColor: theme.body_background},
+          { backgroundColor: theme.body_background },
         ]}>
         <Icon width="37" height="38" name="profile" />
         <View style={style.name_phone_container}>
-          <Text style={[style.name_device, {color: theme.device_list_title}]}>
+          <Text style={[style.name_device, { color: theme.device_list_title }]}>
             {item.name}
           </Text>
           <Text
-            style={[style.phone_device, {color: theme.device_list_subtitle}]}>
+            style={[style.phone_device, { color: theme.device_list_subtitle }]}>
             {item.phoneNumber}
           </Text>
         </View>
@@ -37,10 +83,88 @@ const ListItem = ({item, theme, navigation, deleteDevice}) => {
         <TouchableOpacity onPress={handleDeleteDevice} style={style.icon}>
           <Icon width={width} height={height} name="delete" />
         </TouchableOpacity>
-        <TouchableOpacity style={style.icon}>
+        <TouchableOpacity onPress={toggleOverlay} style={style.icon}>
           <Icon width={width} height={height} name="edit" />
         </TouchableOpacity>
       </View>
+
+      {/*****************Ovarlay Edit******************* */}
+      <View
+        style={{
+          backgroundColor: theme.device_add_background,
+        }}>
+
+        <Overlay
+          overlayStyle={[
+            styles.Overlay,
+            { backgroundColor: theme.overlay_background },
+          ]}
+          isVisible={state.visible}
+          onBackdropPress={toggleOverlay}>
+
+          <View style={[styles.container]}>
+            <Text style={[styles.editTitle, { color: theme.overlay_title }]}>{screen.edit}</Text>
+            <View style={styles.container_form}>
+              <Text style={[styles.label, { color: theme.overlay_title }]}>
+                {screen.name_label}
+              </Text>
+              <Input
+                inputContainerStyle={styles.input_container_style}
+                containerStyle={styles.input_container}
+                placeholder={item.name}
+                inputStyle={{ color: theme.overlay_title }}
+                onChangeText={(text) => {
+                  set({
+                    ...state,
+                    nameEdit: text,
+                  });
+                }}
+              />
+            </View>
+            <View style={styles.container_form}>
+              <Text style={[styles.label, { color: theme.overlay_title }]}>
+                {screen.cel_label}
+              </Text>
+              <Input
+                keyboardType="numeric"
+                inputContainerStyle={styles.input_container_style}
+                containerStyle={styles.input_container}
+                placeholder={item.phoneNumber}
+                inputStyle={{ color: theme.overlay_title }}
+                onChangeText={(text) => {
+                  set({
+                    ...state,
+                    phoneEdit: text,
+                  });
+                }}
+                maxLength={10}
+              />
+            </View>
+            <View style={styles.container_buttons}>
+              <Button
+                title={screen.add_cancel_label}
+                buttonStyle={[
+                  styles.button_cancel,
+                  { backgroundColor: theme.overlay_button_regular },
+                ]}
+                onPress={toggleOverlay}
+                titleStyle={{ color: theme.overlay_button_title }}
+              />
+              <Button
+                title={screen.add_confirm_label}
+                buttonStyle={[
+                  styles.button_confirm,
+                  { backgroundColor: theme.overlay_button_primary },
+                ]}
+                
+                titleStyle={{ color: theme.overlay_button_title }}
+                onPress={handleEditDevice}
+              />
+            </View>
+          </View>
+        </Overlay>
+      </View>
+
     </View>
   );
 };
@@ -82,7 +206,57 @@ const style = StyleSheet.create({
     paddingHorizontal: 7,
   },
 });
+
+const styles = StyleSheet.create({
+  Overlay: {
+    height: 350,
+    borderRadius: 5,
+  },
+  container: {
+    flex: 1,
+    width: '80%',
+    paddingVertical: 30,
+    paddingHorizontal: 10,
+  },
+  editTitle: {
+    fontFamily: 'Roboto_Regular',
+    fontSize: 25,
+    fontWeight: 'bold'
+  },
+  container_form: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginVertical: 5,
+  },
+  container_buttons: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+  },
+  label: {
+    fontFamily: 'Roboto_Regular',
+    fontSize: 17,
+  },
+  button_cancel: {
+    borderRadius: 5,
+    paddingHorizontal: 10,
+  },
+  button_confirm: {
+    borderRadius: 5,
+    paddingHorizontal: 20,
+    marginLeft: 15,
+  },
+  input_container_style: {
+    marginTop: 20,
+  },
+  input_container: {
+    width: '70%',
+  },
+});
+
+
+
 const mapDispatchToProps = {
   deleteDevice,
+  editDevice
 };
 export default connect(null, mapDispatchToProps)(ListItem);
